Split package into lib and sys
