+++ sdl2-sys/src/lib.rs
@@ -1,0 +1,26 @@
+#![allow(non_camel_case_types)]
+
+extern crate libc;
+
+pub mod audio;
+pub mod clipboard;
+pub mod controller;
+pub mod cpuinfo;
+pub mod event;
+pub mod filesystem;
+pub mod haptic;
+pub mod gesture;
+pub mod joystick;
+pub mod keyboard;
+pub mod messagebox;
+pub mod rect;
+pub mod pixels;
+pub mod render;
+pub mod rwops;
+pub mod surface;
+pub mod touch;
+pub mod video;
+pub mod mouse;
+pub mod sdl;
+pub mod timer;
+pub mod version;
