+++ src/sdl2/clipboard.rs
@@ -1,18 +1,7 @@
 use SdlResult;
 use get_error;
 
-#[allow(non_camel_case_types)]
-pub mod ll {
-    use libc::{c_int, c_char};
-
-    pub type SDL_bool = c_int;
-
-    extern "C" {
-        pub fn SDL_SetClipboardText(text: *const c_char) -> c_int;
-        pub fn SDL_GetClipboardText() -> *const c_char;
-        pub fn SDL_HasClipboardText() -> SDL_bool;
-    }
-}
+pub use sys::clipboard as ll;
 
 pub fn set_clipboard_text(text: &String) -> SdlResult<()> {
     unsafe {
