+++ src/sdl2/filesystem.rs
@@ -1,15 +1,7 @@
 use SdlResult;
 use get_error;
 
-#[allow(non_camel_case_types)]
-pub mod ll {
-    use libc::{c_char};
-
-    extern "C" {
-        pub fn SDL_GetBasePath() -> *const c_char;
-        pub fn SDL_GetPrefPath(arg: *const c_char, app: *const c_char) -> *const c_char;
-    }
-}
+pub use sys::filesystem as ll;
 
 pub fn get_base_path() -> SdlResult<String> {
     let result = unsafe {
