+++ src/sdl2/touch.rs
@@ -1,33 +1,9 @@
 use std::ptr;
 
-pub type TouchDevice = ll::SDL_TouchID;
+pub use sys::touch as ll;
 
-#[deriving(PartialEq, Copy)]
-#[repr(C)]
-pub struct Finger {
-    id: TouchDevice,
-    x: f32,
-    y: f32,
-    pressure: f32,
-}
-
-#[allow(non_camel_case_types)]
-pub mod ll {
-    use libc::{c_int, int64_t};
-    use touch::Finger;
-
-    pub type SDL_TouchID = int64_t;
-    pub type SDL_FingerID = int64_t;
-    pub type SDL_Finger = Finger;
-
-    extern "C" {
-        pub fn SDL_GetNumTouchDevices() -> c_int;
-        pub fn SDL_GetTouchDevice(index: c_int) -> SDL_TouchID;
-        pub fn SDL_GetNumTouchFingers(touchID: SDL_TouchID) -> c_int;
-        pub fn SDL_GetTouchFinger(touchID: SDL_TouchID, index: c_int) ->
-                  *const SDL_Finger;
-    }
-}
+pub type Finger = ll::Finger;
+pub type TouchDevice = ll::TouchDevice;
 
 pub fn get_num_touch_devices() -> int {
     unsafe { ll::SDL_GetNumTouchDevices() as int }
