+++ src/sdl2/version.rs
@@ -6,23 +6,7 @@ use std::fmt;
 use std::c_str::CString;
 use std::borrow::ToOwned;
 
-#[doc(hidden)]
-#[allow(non_camel_case_types)]
-pub mod ll {
-    use libc::{uint8_t, c_char, c_int};
-    #[deriving(Copy, Clone)]
-    #[repr(C)]
-    pub struct SDL_version {
-        pub major: uint8_t,
-        pub minor: uint8_t,
-        pub patch: uint8_t,
-    }
-    extern "C" {
-        pub fn SDL_GetVersion(ver: *mut SDL_version);
-        pub fn SDL_GetRevision() -> *const c_char;
-        pub fn SDL_GetRevisionNumber() -> c_int;
-    }
-}
+pub use sys::version as ll;
 
 /// A structure that contains information about the version of SDL in use.
 #[deriving(PartialEq, Copy, Clone)]
